+++ NEWS.md
@@ -1,5 +1,10 @@
 # inlabru (development version)
 
+## Bug fixes
+
+* Fix logic for `bru_set_missing<inla.surv>()` to correctly handle the
+  `inla.surv()` class variations (version `2.13.0.9002`)
+
 # inlabru 2.13.0
 
 ## New features